Refactor: Extracted playlist functions into extension methods
