+++ src/Commands/PlaylistClear.cs
@@ -5,20 +5,22 @@
 
 using Media.Infrastructure;
 using Media.Infrastructure.BaseCommands;
+using Media.Interop;
 
 namespace Media.Commands;
 
 [Example("Clear a playlist", "media playlist clear -p test.m3u")]
-internal sealed class PlaylistClear : BasePlaylistCommand<BasePlalistSettings>
+internal sealed class PlaylistClear : BaseFileWorkCommand<BasePlalistSettings>
 {
     public override async Task<int> ExecuteAsync(CommandContext context, BasePlalistSettings settings)
     {
-        var list = await LoadFromFile(settings.PlaylistName);
+        var list = new List<string>();
+        await list.LoadFromFile(settings.PlaylistName);
 
         int count = list.Count;
         list.Clear();
 
-        await SaveToFile(list, settings.PlaylistName, false);
+        await list.SaveToFile(settings.PlaylistName, false);
 
         Terminal.GreenText($"Removed {count} files from playlist {settings.PlaylistName}");
